remove CLIP-analysis, refactor code
