--- clipper/src/input_norm.py
@@ -1,1 +1,0 @@
-__author__ = 'gpratt'
