+++ clipper/src/peak_curve.py
@@ -1,0 +1,13 @@
+'''
+Created on Jul 25, 2012
+@author: mlovci
+@author: gabrielp
+
+Refactored on June 29, 2020
+@author: algaebrown
+
+'''
+
+############################################################
+####### This files is for peak curve fitting ###############
+############################################################
