--- clipper/test/test_CLIP_Analysis_Display.py
@@ -1,18 +1,0 @@
-'''
-Created on Sep 18, 2012
-
-@author: gabrielp
-'''
-import unittest
-
-
-class Test(unittest.TestCase):
-
-
-    def testName(self):
-        pass
-
-
-if __name__ == "__main__":
-    #import sys;sys.argv = ['', 'Test.testName']
-    unittest.main()
