+++ setup.py
@@ -39,8 +39,8 @@ setup(
     
     entry_points = {
                     'console_scripts': [
-                                        'clipper = clipper.src.peakfinder:call_main',
-                                        'clip_analysis = clipper.src.CLIP_analysis_runner:call_main',],
+                                        'clipper = clipper.src.main:call_main',
+                                        ],
                     },
 
     #metadata for upload to PyPI
